TFP-3997 egen tjeneste i uttak for å sjekke om relevante arbeidsforhold er endret
